requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
werkzeug==3.0.1
flask-login==0.6.3
flask-cors==4.0.0
requests
python-dotenv
